Refund table

+++ src/main/java/com/myweb/mamababy/dtos/PaymentDTO.java
@@ -17,7 +17,7 @@ public class PaymentDTO {
 
     @Min(value = 5000, message = "Price must be greater than or equal to 0")
     @Max(value = 10000000, message = "Price must be less than or equal to 10,000,000")
-    private int finalAmount;
+    private float finalAmount;
 
     @NotEmpty(message = "Bank code is required")
     private String bankCode;
